+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Added management of `ISkinSensor`. (https://github.com/robotology/wearables/pull/139)
 ## [1.2.2] - 2021-10-28
 
 ### Removed